requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
Pillow==9.5.0
requests==2.30.0
urllib3==2.0.2
